backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.0.1
google-genai>=0.3.0
